refactor: move core interfaces into their respective folders
