+++ src/Iterable/Filtered.php
@@ -8,8 +8,6 @@ declare(strict_types=1);
 
 namespace Mono\Iterable;
 
-use Mono\Sequence;
-
 /**
  * Filters items in a {@see Sequence} by a predicate.
  *
