+++ src/Iterable/Mapped.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 namespace Mono\Iterable;
 
 use Closure;
-use Mono\Sequence;
 
 /**
  * A sequence where each item is the result of applying a function to the original items.
