+++ src/Iterable/SequenceOf.php
@@ -10,7 +10,6 @@ namespace Mono\Iterable;
 
 use ArrayIterator;
 use Iterator;
-use Mono\Sequence;
 
 /**
  * A sequence backed by a PHP array.
