+++ src/Logic/IsUrl.php
@@ -8,8 +8,7 @@ declare(strict_types=1);
 
 namespace Mono\Logic;
 
-use Mono\Logic;
-use Mono\Text;
+use Mono\Text\Text;
 
 /**
  * {@see Logic} that returns true if a {@see Text} is a valid URL.
