+++ src/Logic/IsUuid.php
@@ -8,8 +8,6 @@ declare(strict_types=1);
 
 namespace Mono\Logic;
 
-use Mono\Logic;
-
 /**
  * {@see Logic} that returns true if the text is a valid UUID v1–v5.
  *
