+++ src/Logic/Yes.php
@@ -8,8 +8,6 @@ declare(strict_types=1);
 
 namespace Mono\Logic;
 
-use Mono\Logic;
-
 /**
  * {@see Logic} that always returns true.
  *
