+++ src/Text/HtmlSanitized.php
@@ -8,8 +8,6 @@ declare(strict_types=1);
 
 namespace Mono\Text;
 
-use Mono\Text;
-
 /**
  * {@see Text} sanitized for HTML output.
  *
